requirements.txt
aidial-sdk==0.22.1
aidial-client>=0.3.0
mcp>=1.16.0
pydantic>=2.11.10
faiss-cpu>=1.12.0
sentence-transformers>=5.1.1
beautifulsoup4>=4.14.2
PyPDF2>=3.0.1
numpy>=2.3.3
pandas>=2.3.3
tabulate>=0.9.0
langchain>=0.3.27